+++ src/Products/PloneMeeting/tests/testAnnexes.py
@@ -20,6 +20,7 @@ from collective.iconifiedcategory.utils import update_all_categorized_elements
 from imio.actionspanel.interfaces import IContentDeletable
 from imio.annex.columns import ActionsColumn
 from imio.annex.utils import get_annexes_to_print
+from imio.helpers.content import get_vocab
 from imio.helpers.content import get_vocab_values
 from plone import api
 from plone.app.textfield.value import RichTextValue
@@ -42,11 +43,9 @@ from Products.PloneMeeting.utils import get_annexes
 from tempfile import mkdtemp
 from time import sleep
 from zope.annotation import IAnnotations
-from zope.component import queryUtility
 from zope.event import notify
 from zope.interface import Invalid
 from zope.lifecycleevent import ObjectModifiedEvent
-from zope.schema.interfaces import IVocabularyFactory
 
 
 class testAnnexes(PloneMeetingTestCase):
@@ -1417,7 +1416,7 @@ class testAnnexes(PloneMeetingTestCase):
         # get vocabulary name
         type_info = self.portal.portal_types.get(annex_type.portal_type)
         vocab_name = type_info.lookupSchema()['other_mc_correspondences'].value_type.vocabularyName
-        terms = queryUtility(IVocabularyFactory, vocab_name)(annex_type)._terms
+        terms = get_vocab(annex_type, vocab_name)._terms
         self.assertRaises(Invalid, other_mc_correspondences_constraint, [terms[0].value, terms[1].value])
         self.assertTrue(other_mc_correspondences_constraint([terms[0].value]))
         self.assertTrue(other_mc_correspondences_constraint([terms[1].value]))
@@ -1430,7 +1429,6 @@ class testAnnexes(PloneMeetingTestCase):
         # get vocabulary name
         type_info = self.portal.portal_types.get(annex_type.portal_type)
         vocab_name = type_info.lookupSchema()['other_mc_correspondences'].value_type.vocabularyName
-        vocab = queryUtility(IVocabularyFactory, vocab_name)
         # build expected result depending on existing MC
         expected = []
         for mc in self.tool.objectValues('MeetingConfig'):
@@ -1449,12 +1447,14 @@ class testAnnexes(PloneMeetingTestCase):
                 u'{0} ➔ Item annexes ➔ Other annex(es)'.format(mc_title),
                 u'{0} ➔ Item decision annexes ➔ Decision annex(es)'.format(mc_title)]
             expected.extend(values)
-        self.assertEqual([term.title for term in vocab(annex_type)._terms], expected)
+        self.assertEqual(
+            [term.title for term in get_vocab(annex_type, vocab_name)._terms],
+            expected)
 
     def test_pm_Annex_type_only_for_meeting_managers(self):
         """An ItemAnnexContentCategory may be defined only selectable by MeetingManagers."""
         cfg = self.meetingConfig
-        vocab = queryUtility(IVocabularyFactory, 'collective.iconifiedcategory.categories')
+        vocab = get_vocab(None, 'collective.iconifiedcategory.categories', only_factory=True)
         self.changeUser('pmCreator1')
         item = self.create('MeetingItem')
         annex = self.addAnnex(item)